// backend/src/modules/catalog/enums/fulfillment.enum.ts
export enum WeightUnit {
  KG = 'KG',
  G = 'G',
  LB = 'LB',
  OZ = 'OZ',
}

export enum DimensionUnit {
  CM = 'CM',
  M = 'M',
  IN = 'IN',
}

export enum DigitalDeliveryType {
  DOWNLOAD = 'DOWNLOAD',
  ACCESS_LINK = 'ACCESS_LINK',
  LICENSE_KEY = 'LICENSE_KEY',
  EXTERNAL = 'EXTERNAL',
}

export enum ServiceDeliveryType {
  ONLINE = 'ONLINE',
  ONSITE = 'ONSITE',
  LOCATION = 'LOCATION',
}

export enum ServiceDurationUnit {
  MINUTES = 'MINUTES',
  HOURS = 'HOURS',
  DAYS = 'DAYS',
}